iter12 fix убрал лишний файл
